Admin user functionality was added.
'is_admin' was added in the User.php class casts function.
PostPolicy was updated with the before function.
The PostPolicy was defined in the AppServiceProvider.class
A migration was created to add 'is_admin' column to users table.
Default DatabaseSeeder was updated to create the admin user.
Route was created for the '/admin' route.