Admin CRUD feature
AdminPostController created with edit, update and destroy method.
IsAdminMiddleware was created to check the user was authenticated and the user has the admin role.
index blade file was created inside the admin\posts directory.
The edit blade file also created.
The admin routes was created as a group enclosed by the 'is-admin' middleware.